+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update @mia-platform/lc39 2.1.1 -> 2.1.2
 - Update @types/node 12.0.0 -> 12.0.2
 - Update eslint configuration
+- Add extra headers proxy during internal http calls
 
 ### v1.0.0-rc.2 (2019-05-08)